feat(iast): include as iast telemetry metrics tags in the root span of the trace
