+++ ddtrace/appsec/_iast/taint_sinks/command_injection.py
@@ -14,8 +14,8 @@ from ddtrace.contrib import trace_utils
 from ddtrace.internal.logger import get_logger
 
 from .. import oce
-from ..._asm_request_context import increment_iast_span_metric
 from ..._constants import IAST_SPAN_TAGS
+from .._metrics import increment_iast_span_metric
 from .._utils import _has_to_scrub
 from .._utils import _scrub
 from .._utils import _scrub_get_tokens_positions
