+++ ddtrace/appsec/_iast/taint_sinks/path_traversal.py
@@ -3,9 +3,9 @@ from typing import Any
 from ddtrace.internal.logger import get_logger
 
 from .. import oce
-from ..._asm_request_context import increment_iast_span_metric
 from ..._constants import IAST_SPAN_TAGS
 from .._metrics import _set_metric_iast_instrumented_sink
+from .._metrics import increment_iast_span_metric
 from .._patch import set_and_check_module_is_patched
 from .._patch import set_module_unpatched
 from ..constants import EVIDENCE_PATH_TRAVERSAL
