+++ ddtrace/appsec/_iast/taint_sinks/ssrf.py
@@ -7,8 +7,8 @@ from ddtrace.internal.logger import get_logger
 from ddtrace.settings import _config
 
 from .. import oce
-from ..._asm_request_context import increment_iast_span_metric
 from ..._constants import IAST_SPAN_TAGS
+from .._metrics import increment_iast_span_metric
 from .._taint_tracking import taint_ranges_as_evidence_info
 from .._utils import _has_to_scrub
 from .._utils import _scrub
